+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Added `review_status` to tasks, added a DB trigger to update this field when children task reviews change [#5478](https://github.com/raster-foundry/raster-foundry/pull/5478)
+
 ### Changed
 
 - Moved task grid creation into a background task [#5476](https://github.com/raster-foundry/raster-foundry/pull/5476)